Move ClientMsgHandler to the actual handlers package. This eliminates having two handlers packages and hints that the sub packages will be referencing it.
