+++ src/main/java/in/dragonbra/javasteam/steam/steamclient/SteamClient.java
@@ -2,7 +2,7 @@ package in.dragonbra.javasteam.steam.steamclient;
 
 import in.dragonbra.javasteam.base.ClientMsgProtobuf;
 import in.dragonbra.javasteam.base.IPacketMsg;
-import in.dragonbra.javasteam.handlers.ClientMsgHandler;
+import in.dragonbra.javasteam.steam.handlers.ClientMsgHandler;
 import in.dragonbra.javasteam.protobufs.steamclient.SteammessagesClientserver.CMsgClientCMList;
 import in.dragonbra.javasteam.steam.CMClient;
 import in.dragonbra.javasteam.steam.handlers.steamapps.SteamApps;
